requirements.txt
asn1crypto==0.22.0
backports.ssl-match-hostname==3.5.0.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==2.0.3
docker==2.5.1
docker-pycreds==0.2.1
enum34==1.1.6
Flask==0.12.2
Flask-Hookserver==1.1.0
gitdb2==2.0.2
GitPython==2.1.5
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pkg-resources==0.0.0
pycparser==2.18
pyOpenSSL==17.2.0
requests==2.18.4
six==1.10.0
smmap2==2.0.3
urllib3==1.22
websocket-client==0.44.0
Werkzeug==0.12.2
